+++ src/routes/ForumPost/BackButton.jsx
@@ -1,7 +1,9 @@
 import React from 'react'
+import { Link } from 'react-router-dom'
 import { makeStyles } from '@material-ui/core/styles';
 import Button from '@material-ui/core/Button';
 import Grid from '@material-ui/core/Grid';
+import { FORUM } from 'constants/routes';
 
 const useStyles = makeStyles(() => ({
   wrapper: {
@@ -18,12 +20,14 @@ const BackButton = () => {
   return (
     <div className={classes.wrapper}>
       <Grid container alignItems="flex-start" justify="flex-start" direction="row">
-        <Button 
-          variant="contained" 
-          className={classes.button}
-        >
-          Back
+        <Link to={FORUM}>
+          <Button
+            variant="contained"
+            className={classes.button}
+          >
+            Back
         </Button>
+        </Link>
       </Grid>
     </div>
   )
