+++ src/routes/ForumPost/CommentData.jsx
@@ -1,0 +1,12 @@
+export const comments = [
+  {
+    id: '1',
+    question: 'Lorem ipsum dolor sit amet, consectetur adipiscing elit. Vivamus eget turpis vulputate, mattis turpis vel, eleifend turpis. Suspendisse non eros magna. Nam viverra, nisi id venenatis facilisis, neque tortor rhoncus eros, et scelerisque lorem est sit amet odio.',
+    date: '13 Feb 2020',
+  },
+  {
+    id: '2',
+    question: 'Lorem ipsum dolor sit amet, consectetur adipiscing elit. Vivamus eget turpis vulputate, mattis turpis vel, eleifend turpis. Suspendisse non eros magna. Nam viverra, nisi id venenatis facilisis, neque tortor rhoncus eros, et scelerisque lorem est sit amet odio.',
+    date: '1 Feb 2020',
+  },
+]
