+++ src/routes/ForumPost/ForumPost.jsx
@@ -1,24 +1,48 @@
 import React from 'react';
 import { makeStyles } from '@material-ui/core/styles';
-import Post from './Post';
-import { ForumPostData } from './ForumPostData';
+import CssBaseline from '@material-ui/core/CssBaseline';
+import Container from '@material-ui/core/Container';
+import Question from './Question';
+import Comments from './Comments';
+import BackButton from './BackButton';
+import ForumLogo from 'assets/Forum/ForumLogo.svg';
 
-const useStyles = makeStyles(() => ({
-  forumList: {
-    marginTop: '10px',
-  }
+const useStyles = makeStyles(theme => ({
+  paper: {
+    marginTop: theme.spacing(10),
+  },
+  header: {
+    fontWeight: 'bold',
+    fontSize: '30px',
+    color: 'teal',
+    textAlign: 'center',
+  },
+  forumLogo: {
+    display: 'block',
+    margin: '0px auto',
+    width: '150px',
+  },
+  forumPostWrapper: {
+    paddingTop: '10px',
+  },
 }));
-const ForumList = () => {
-  const classes = useStyles();
+
+const Forum = () => {
+  const classes = useStyles()
   return (
-    <div className={classes.forumList}>
-      {
-        ForumPostData.map(post => (
-          <Post author={post.author} id={post.id} date={post.date} question={post.question} key={post.id}/>
-        ))
-      }
-    </div>
-  );
+    <Container component="main" maxWidth="sm">
+      <CssBaseline />
+      <div className={classes.paper}>
+        <div className={classes.header}>Community Forum</div>
+        <img src={ForumLogo} alt="" className={classes.forumLogo} />
+        <div className={classes.forumPostWrapper}>
+          <BackButton/>
+          <Question question="Can my child with Autism attend school?" date="12 Feb 2020"/>
+          <Comments />
+        </div>
+      </div>
+    </Container>
+  )
 }
 
-export default ForumList;
+export default Forum;
