--- src/routes/ForumPost/ForumPostData.jsx
@@ -1,14 +1,0 @@
-export const ForumPostData = [
-  {
-    id: '1',
-    question: 'Can my child with Autism attend school?',
-    date: '13 Feb 2020',
-    author: 'Anonymous',
-  },
-  {
-    id: '2',
-    question: 'What to do after diagnosis of Autism?',
-    date: '1 Feb 2020',
-    author: 'Misaka Mikoto',
-  },
-]
